fixup! internal/credential: add pagination to libraries
