+++ internal/credential/vault/service_list_ext_test.go
@@ -17,6 +17,36 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+type fakeWriter struct {
+	db.Writer
+}
+
+func TestNewLibraryListingService(t *testing.T) {
+	t.Parallel()
+	ctx := context.Background()
+	t.Run("success", func(t *testing.T) {
+		t.Parallel()
+		got, err := vault.NewLibraryListingService(ctx, &fakeWriter{}, &vault.Repository{})
+		require.NoError(t, err)
+		require.NotNil(t, got)
+	})
+	t.Run("nil-writer", func(t *testing.T) {
+		t.Parallel()
+		_, err := vault.NewLibraryListingService(ctx, nil, &vault.Repository{})
+		require.Error(t, err)
+	})
+	t.Run("nil-interface-writer", func(t *testing.T) {
+		t.Parallel()
+		_, err := vault.NewLibraryListingService(ctx, (*fakeWriter)(nil), &vault.Repository{})
+		require.Error(t, err)
+	})
+	t.Run("nil-repo", func(t *testing.T) {
+		t.Parallel()
+		_, err := vault.NewLibraryListingService(ctx, &fakeWriter{}, nil)
+		require.Error(t, err)
+	})
+}
+
 func TestLibraryListingService_List(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
